use new code on tagged articles feed
