+++ app/javascript/packs/feedNavigation.js
@@ -1,0 +1,6 @@
+import('../feedNavigation/feedNavigation').then(
+  ({ initializeFeedOptionsDropdown, initializeMobileNavigationDrawer }) => {
+    initializeFeedOptionsDropdown();
+    initializeMobileNavigationDrawer();
+  },
+);
